+++ src/lib/createables.ts
@@ -1,0 +1,638 @@
+import { Bank } from './types';
+import itemID from './util/itemID';
+
+interface Createable {
+	name: string;
+	outputItems: Bank;
+	inputItems: Bank;
+	smithingLevel?: number;
+	addOutputToCollectionLog?: boolean;
+	cantHaveItems: Bank;
+	firemakingLevel?: number;
+	craftingLevel?: number;
+	prayerLevel?: number;
+}
+
+const Createables: Createable[] = [
+	{
+		name: 'Godsword blade',
+		inputItems: {
+			[itemID('Godsword shard 1')]: 1,
+			[itemID('Godsword shard 2')]: 1,
+			[itemID('Godsword shard 3')]: 1
+		},
+		outputItems: {
+			[itemID('Godsword blade')]: 1
+		},
+		smithingLevel: 80,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Armadyl godsword',
+		inputItems: {
+			[itemID('Godsword blade')]: 1,
+			[itemID('Armadyl hilt')]: 1
+		},
+		outputItems: {
+			[itemID('Armadyl godsword')]: 1
+		},
+		smithingLevel: 80,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Bandos godsword',
+		inputItems: {
+			[itemID('Godsword blade')]: 1,
+			[itemID('Bandos hilt')]: 1
+		},
+		outputItems: {
+			[itemID('Bandos godsword')]: 1
+		},
+		smithingLevel: 80,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Saradomin godsword',
+		inputItems: {
+			[itemID('Godsword blade')]: 1,
+			[itemID('Saradomin hilt')]: 1
+		},
+		outputItems: {
+			[itemID('Saradomin godsword')]: 1
+		},
+		smithingLevel: 80,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Zamorak godsword',
+		inputItems: {
+			[itemID('Godsword blade')]: 1,
+			[itemID('Zamorak hilt')]: 1
+		},
+		outputItems: {
+			[itemID('Zamorak godsword')]: 1
+		},
+		smithingLevel: 80,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Dragonfire shield',
+		inputItems: {
+			[itemID('Draconic visage')]: 1,
+			[itemID('Anti-dragon shield')]: 1
+		},
+		outputItems: {
+			// Uncharged dragonfire shield
+			11284: 1
+		},
+		smithingLevel: 90,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Dragonfire ward',
+		inputItems: {
+			[itemID('Skeletal visage')]: 1,
+			[itemID('Anti-dragon shield')]: 1
+		},
+		outputItems: {
+			// Uncharged Dragonfire ward
+			22003: 1
+		},
+		smithingLevel: 90,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Infernal pickaxe',
+		inputItems: {
+			[itemID('Dragon pickaxe')]: 1,
+			[itemID('Smouldering stone')]: 1
+		},
+		outputItems: {
+			[itemID('Infernal pickaxe')]: 1
+		},
+		smithingLevel: 85,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Malediction ward',
+		inputItems: {
+			[itemID('Malediction shard 1')]: 1,
+			[itemID('Malediction shard 2')]: 1,
+			[itemID('Malediction shard 3')]: 1
+		},
+		outputItems: {
+			[itemID('Malediction ward')]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Odium ward',
+		inputItems: {
+			[itemID('Odium shard 1')]: 1,
+			[itemID('Odium shard 2')]: 1,
+			[itemID('Odium shard 3')]: 1
+		},
+		outputItems: {
+			[itemID('Odium ward')]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Crystal key',
+		inputItems: {
+			[itemID('Loop half of key')]: 1,
+			[itemID('Tooth half of key')]: 1
+		},
+		outputItems: {
+			[itemID('Crystal key')]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Verac's armour set",
+		inputItems: {
+			[itemID("Verac's helm")]: 1,
+			[itemID("Verac's brassard")]: 1,
+			[itemID("Verac's plateskirt")]: 1,
+			[itemID("Verac's flail")]: 1
+		},
+		outputItems: {
+			[itemID("Verac's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Veracs',
+		inputItems: {
+			[itemID("Verac's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Verac's helm")]: 1,
+			[itemID("Verac's brassard")]: 1,
+			[itemID("Verac's plateskirt")]: 1,
+			[itemID("Verac's flail")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Dharok's armour set",
+		inputItems: {
+			[itemID("Dharok's helm")]: 1,
+			[itemID("Dharok's platebody")]: 1,
+			[itemID("Dharok's platelegs")]: 1,
+			[itemID("Dharok's greataxe")]: 1
+		},
+		outputItems: {
+			[itemID("Dharok's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Dharoks',
+		inputItems: {
+			[itemID("Dharok's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Dharok's helm")]: 1,
+			[itemID("Dharok's platebody")]: 1,
+			[itemID("Dharok's platelegs")]: 1,
+			[itemID("Dharok's greataxe")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Guthan's armour set",
+		inputItems: {
+			[itemID("Guthan's helm")]: 1,
+			[itemID("Guthan's platebody")]: 1,
+			[itemID("Guthan's chainskirt")]: 1,
+			[itemID("Guthan's warspear")]: 1
+		},
+		outputItems: {
+			[itemID("Guthan's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Guthans',
+		inputItems: {
+			[itemID("Guthan's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Guthan's helm")]: 1,
+			[itemID("Guthan's platebody")]: 1,
+			[itemID("Guthan's chainskirt")]: 1,
+			[itemID("Guthan's warspear")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Ahrim's armour set",
+		inputItems: {
+			[itemID("Ahrim's hood")]: 1,
+			[itemID("Ahrim's robetop")]: 1,
+			[itemID("Ahrim's robeskirt")]: 1,
+			[itemID("Ahrim's staff")]: 1
+		},
+		outputItems: {
+			[itemID("Ahrim's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Ahrims',
+		inputItems: {
+			[itemID("Ahrim's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Ahrim's hood")]: 1,
+			[itemID("Ahrim's robetop")]: 1,
+			[itemID("Ahrim's robeskirt")]: 1,
+			[itemID("Ahrim's staff")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Torag's armour set",
+		inputItems: {
+			[itemID("Torag's helm")]: 1,
+			[itemID("Torag's platebody")]: 1,
+			[itemID("Torag's platelegs")]: 1,
+			[itemID("Torag's hammers")]: 1
+		},
+		outputItems: {
+			[itemID("Torag's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Torags',
+		inputItems: {
+			[itemID("Torag's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Torag's helm")]: 1,
+			[itemID("Torag's platebody")]: 1,
+			[itemID("Torag's platelegs")]: 1,
+			[itemID("Torag's hammers")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: "Karil's armour set",
+		inputItems: {
+			[itemID("Karil's coif")]: 1,
+			[itemID("Karil's leathertop")]: 1,
+			[itemID("Karil's leatherskirt")]: 1,
+			[itemID("Karil's crossbow")]: 1
+		},
+		outputItems: {
+			[itemID("Karil's armour set")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Karils',
+		inputItems: {
+			[itemID("Karil's armour set")]: 1
+		},
+		outputItems: {
+			[itemID("Karil's coif")]: 1,
+			[itemID("Karil's leathertop")]: 1,
+			[itemID("Karil's leatherskirt")]: 1,
+			[itemID("Karil's crossbow")]: 1
+		},
+		smithingLevel: 1,
+		cantHaveItems: {}
+	},
+	/**
+	 * Prospector outfit
+	 */
+	{
+		name: 'Prospector helmet',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Prospector helmet')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 40
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Prospector jacket',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Prospector jacket')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 60
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Prospector legs',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Prospector legs')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 50
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Prospector boots',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Prospector boots')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 30
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Mining gloves',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Unidentified minerals')]: 60
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Superior mining gloves',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Superior mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Unidentified minerals')]: 120
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Expert mining gloves',
+		addOutputToCollectionLog: true,
+		outputItems: {
+			[itemID('Expert mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Superior mining gloves')]: 1,
+			[itemID('Mining gloves')]: 1,
+			[itemID('Unidentified minerals')]: 60
+		},
+		cantHaveItems: {}
+	},
+	{
+		name: 'Master clue',
+		inputItems: {
+			[itemID('Clue scroll (easy)')]: 1,
+			[itemID('Clue scroll (medium)')]: 1,
+			[itemID('Clue scroll (hard)')]: 1,
+			[itemID('Clue scroll (elite)')]: 1
+		},
+		outputItems: {
+			[itemID('Clue scroll (master)')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Clue scroll (master)')]: 1
+		}
+	},
+	{
+		name: 'Infernal axe',
+		inputItems: {
+			[itemID('Dragon axe')]: 1,
+			[itemID('Smouldering stone')]: 1
+		},
+		outputItems: {
+			[itemID('Infernal axe')]: 1
+		},
+		firemakingLevel: 85,
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful',
+		inputItems: {
+			[itemID('Mark of grace')]: 260
+		},
+		outputItems: {
+			[itemID('Graceful hood')]: 1,
+			[itemID('Graceful top')]: 1,
+			[itemID('Graceful legs')]: 1,
+			[itemID('Graceful gloves')]: 1,
+			[itemID('Graceful boots')]: 1,
+			[itemID('Graceful cape')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful hood',
+		inputItems: {
+			[itemID('Mark of grace')]: 35
+		},
+		outputItems: {
+			[itemID('Graceful hood')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful top',
+		inputItems: {
+			[itemID('Mark of grace')]: 55
+		},
+		outputItems: {
+			[itemID('Graceful top')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful legs',
+		inputItems: {
+			[itemID('Mark of grace')]: 60
+		},
+		outputItems: {
+			[itemID('Graceful legs')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful gloves',
+		inputItems: {
+			[itemID('Mark of grace')]: 30
+		},
+		outputItems: {
+			[itemID('Graceful gloves')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful boots',
+		inputItems: {
+			[itemID('Mark of grace')]: 40
+		},
+		outputItems: {
+			[itemID('Graceful boots')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Graceful cape',
+		inputItems: {
+			[itemID('Mark of grace')]: 40
+		},
+		outputItems: {
+			[itemID('Graceful cape')]: 1
+		},
+		cantHaveItems: {},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Hell cat ears',
+		inputItems: {
+			[itemID('Cat ears')]: 1,
+			[itemID('Red dye')]: 1
+		},
+		outputItems: {
+			[itemID('Hell cat ears')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Hell cat ears')]: 1
+		},
+		addOutputToCollectionLog: true
+	},
+	/** Runecrafting Pouches */
+	{
+		name: 'Small pouch',
+		inputItems: {
+			[itemID('Leather')]: 10
+		},
+		outputItems: {
+			[itemID('Small pouch')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Small pouch')]: 1
+		},
+		addOutputToCollectionLog: true
+	},
+	{
+		name: 'Medium pouch',
+		inputItems: {
+			[itemID('Leather')]: 20
+		},
+		outputItems: {
+			[itemID('Medium pouch')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Medium pouch')]: 1
+		},
+		addOutputToCollectionLog: true,
+		craftingLevel: 10
+	},
+	{
+		name: 'Large pouch',
+		inputItems: {
+			[itemID('Leather')]: 30
+		},
+		outputItems: {
+			[itemID('Large pouch')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Large pouch')]: 1
+		},
+		addOutputToCollectionLog: true,
+		craftingLevel: 20
+	},
+	{
+		name: 'Giant pouch',
+		inputItems: {
+			[itemID('Leather')]: 40
+		},
+		outputItems: {
+			[itemID('Giant pouch')]: 1
+		},
+		cantHaveItems: {
+			[itemID('Giant pouch')]: 1
+		},
+		addOutputToCollectionLog: true,
+		craftingLevel: 30
+	},
+	// Spirit Shields
+	{
+		name: 'Blessed spirit shield',
+		inputItems: {
+			[itemID('Spirit shield')]: 1,
+			[itemID('Holy elixir')]: 1
+		},
+		outputItems: {
+			[itemID('Blessed spirit shield')]: 1
+		},
+		prayerLevel: 85,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Spectral spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Spectral sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Spectral spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Arcane spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Arcane sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Arcane spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85,
+		cantHaveItems: {}
+	},
+	{
+		name: 'Elysian spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Elysian sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Elysian spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85,
+		cantHaveItems: {}
+	}
+];
+
+export default Createables;
