--- src/lib/createables/crafting.ts
@@ -1,73 +1,0 @@
-import itemID from '../util/itemID';
-import { Bank } from '../types';
-
-interface Createable {
-	name: string;
-	outputItems: Bank;
-	inputItems: Bank;
-	smithingLevel?: number;
-	addOutputToCollectionLog?: boolean;
-	cantHaveItems: Bank;
-	firemakingLevel?: number;
-	craftingLevel?: number;
-	prayerLevel?: number;
-}
-const Crafting: Createable[] = [
-	{
-		name: 'Small pouch',
-		inputItems: {
-			[itemID('Leather')]: 10
-		},
-		outputItems: {
-			[itemID('Small pouch')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Small pouch')]: 1
-		},
-		addOutputToCollectionLog: true
-	},
-	{
-		name: 'Medium pouch',
-		inputItems: {
-			[itemID('Leather')]: 20
-		},
-		outputItems: {
-			[itemID('Medium pouch')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Medium pouch')]: 1
-		},
-		addOutputToCollectionLog: true,
-		craftingLevel: 10
-	},
-	{
-		name: 'Large pouch',
-		inputItems: {
-			[itemID('Leather')]: 30
-		},
-		outputItems: {
-			[itemID('Large pouch')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Large pouch')]: 1
-		},
-		addOutputToCollectionLog: true,
-		craftingLevel: 20
-	},
-	{
-		name: 'Giant pouch',
-		inputItems: {
-			[itemID('Leather')]: 40
-		},
-		outputItems: {
-			[itemID('Giant pouch')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Giant pouch')]: 1
-		},
-		addOutputToCollectionLog: true,
-		craftingLevel: 30
-	}
-];
-
-export default Crafting;
