--- src/lib/createables/misc.ts
@@ -1,98 +1,0 @@
-import itemID from '../util/itemID';
-import { Bank } from '../types';
-
-interface Createable {
-	name: string;
-	outputItems: Bank;
-	inputItems: Bank;
-	smithingLevel?: number;
-	addOutputToCollectionLog?: boolean;
-	cantHaveItems: Bank;
-	firemakingLevel?: number;
-	craftingLevel?: number;
-	prayerLevel?: number;
-}
-const Misc: Createable[] = [
-	{
-		name: 'Malediction ward',
-		inputItems: {
-			[itemID('Malediction shard 1')]: 1,
-			[itemID('Malediction shard 2')]: 1,
-			[itemID('Malediction shard 3')]: 1
-		},
-		outputItems: {
-			[itemID('Malediction ward')]: 1
-		},
-		smithingLevel: 1,
-		cantHaveItems: {}
-	},
-	{
-		name: 'Odium ward',
-		inputItems: {
-			[itemID('Odium shard 1')]: 1,
-			[itemID('Odium shard 2')]: 1,
-			[itemID('Odium shard 3')]: 1
-		},
-		outputItems: {
-			[itemID('Odium ward')]: 1
-		},
-		smithingLevel: 1,
-		cantHaveItems: {}
-	},
-	{
-		name: 'Crystal key',
-		inputItems: {
-			[itemID('Loop half of key')]: 1,
-			[itemID('Tooth half of key')]: 1
-		},
-		outputItems: {
-			[itemID('Crystal key')]: 1
-		},
-		smithingLevel: 1,
-		cantHaveItems: {}
-	},
-	{
-		name: 'Master clue',
-		inputItems: {
-			[itemID('Clue scroll (easy)')]: 1,
-			[itemID('Clue scroll (medium)')]: 1,
-			[itemID('Clue scroll (hard)')]: 1,
-			[itemID('Clue scroll (elite)')]: 1
-		},
-		outputItems: {
-			[itemID('Clue scroll (master)')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Clue scroll (master)')]: 1
-		}
-	},
-	{
-		name: 'Infernal axe',
-		inputItems: {
-			[itemID('Dragon axe')]: 1,
-			[itemID('Smouldering stone')]: 1
-		},
-		outputItems: {
-			[itemID('Infernal axe')]: 1
-		},
-		firemakingLevel: 85,
-		cantHaveItems: {},
-		addOutputToCollectionLog: true
-	},
-	{
-		name: 'Hell cat ears',
-		inputItems: {
-			[itemID('Cat ears')]: 1,
-			[itemID('Red dye')]: 1
-		},
-		outputItems: {
-			[itemID('Hell cat ears')]: 1
-		},
-		cantHaveItems: {
-			[itemID('Hell cat ears')]: 1
-		},
-		addOutputToCollectionLog: true
-	}
-];
-
-export default Misc;
